ai/pyproject.toml
[project]
name = "warung-ai"
version = "0.1.0"
description = "AI service for Warung AI - menu parsing and conversational analytics"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "boto3>=1.35",
    "langchain-aws~=1.4.5",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.35",
    "openai>=1.109.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.1.0",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "ruff>=0.7",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
